+++ sqlite3_web/extra/js-package/src/utils.ts
@@ -44,6 +44,9 @@ export function serializeError(
   if (error instanceof DOMException && error.name == "AbortError") {
     serializedExceptionType = typeAbortException;
   }
+  if (error instanceof Error) {
+    message += `: ${error.message}`;
+  }
 
   return {
     t: typeErrorResponse,