+++ sqlite3_web/extra/js-package/test/sources.test.ts
@@ -1,0 +1,5 @@
+import * as module from "../src/index";
+
+import { sqliteTestCases } from "./testcases";
+
+sqliteTestCases(module);